Added SVG and propTypes
