+++ .gitignore
@@ -5,9 +5,6 @@
 /.pnp
 .pnp.js
 
-#api key
-.env
-
 # testing
 /coverage
