+++ package.json
@@ -3,11 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
     "axios": "^0.24.0",
+    "bootstrap": "^5.1.3",
     "dotenv": "^10.0.0",
     "env-cmd": "^10.1.0",
+    "prop-types": "^15.8.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-media": "^1.10.0",
     "react-router": "^6.0.2",
     "react-router-dom": "^5.2.0",
