+++ src/Footer.module.css
@@ -1,0 +1,126 @@
+.footerContainer {
+  background: rgba(25, 80, 120, .9);
+  padding: 4rem 0 2rem 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.fa-user-check {
+  margin-left: 0.5rem;
+  font-size: 1.8rem;
+}
+
+.footerLinks {
+  width: 100%;
+  max-width: 1000px;
+  display: flex;
+  justify-content: center;
+}
+
+.footerLinkWrapper {
+  display: flex;
+}
+
+.footerLinkItems {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin: 16px;
+  text-align: left;
+  width: 160px;
+  box-sizing: border-box;
+}
+
+.footerLinkItems h2 {
+  margin-bottom: 16px;
+}
+
+.footerLinkItems > h2 {
+  color: #fff;
+}
+
+.footerLinkItems a {
+  color: #fff;
+  text-decoration: none;
+  margin-bottom: 8px;
+}
+
+.footerLinkItems a:hover {
+  color: #e9e9e9;
+  transition: 0.3s ease-out;
+}
+
+.footerInput::placeholder {
+  color: #b1b1b1;
+}
+
+/* Social Icons */
+.socialIconLink {
+  color: #fff;
+  font-size: 24px;
+}
+
+.socialMedia {
+  max-width: 1000px;
+  width: 100%;
+}
+
+.socialMediaWrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 90%;
+  max-width: 1000px;
+  margin: 40px auto 0 auto;
+}
+
+.socialIcons {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 240px;
+}
+
+.socialLogo {
+  color: #fff;
+  justify-self: start;
+  margin-left: 20px;
+  cursor: pointer;
+  text-decoration: none;
+  font-size: 2rem;
+  display: flex;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.websiteRights {
+  color: #fff;
+  margin-bottom: 16px;
+}
+
+@media screen and (max-width: 820px) {
+  .footerLinks {
+    padding-top: 2rem;
+  }
+
+  .footerInput {
+    width: 100%;
+  }
+
+  .btn {
+    width: 100%;
+  }
+
+  .footerLinkWrapper {
+    flex-direction: column;
+  }
+
+  .socialMediaWrap {
+    flex-direction: column;
+  }
+}
+
+@media screen and (max-width: 768px) {
+}
