--- src/InputWithLabel.module.css
@@ -1,44 +1,0 @@
-/* Styling the label Title in the input */
-.label {
-  color: darkgreen;
-  border-top: 1px solid blue;
-  border-left: 1px solid blue;
-  border-radius: 5px;
-  padding-left: 5px;
-  font-size: 24px;
-  }
-
-/* Styling the input area */
-.input {
-  border: none;
-  border-bottom: 1px solid blue;
-  background-color: transparent;
-  font-size: 24px;
-}
-
-/* Styling the Add buttom in the input section */
-.buttonSubmit {
-  background: rgb(50, 150, 50);
-  border: 1px solid blue;
-  border-radius: 5px;
-  padding: 6px 10px;
-  cursor: pointer;
-}
-
-.buttonSubmit:hover {
-  background-color: rgb(50, 255, 50); 
-}
-
-/* Making input, label and submit button responsiveness */
-@media (max-width: 600px) {
-  .label {
-    color: blue;
-  }
-}
-
-@media (max-width: 600px) {
-  .input {
-    color: whitesmoke;
-    font-size: 12px;
-  }
-}
