--- src/component/TodoListItem.module.css
@@ -1,43 +1,0 @@
-/* Styling the Remove button in the list item section */
-.buttonRemove {
-  background: rgb(150, 50, 50);
-  color: #fff;
-  border: no;
-  border-radius: 5px;
-  padding: 4px 8px;
-  cursor: pointer;
-}
-
-.buttonRemove:hover {
-  background-color: rgb(255, 50, 50); 
-}
-
-/* Styling the list item section */
-li {
-  list-style: none;
-}
-
-.ListItem {
-  display: flex;
-  color: darkgreen;
-  justify-content: space-between;
-  margin-bottom: 20px;
-  background: #83a4d4;
-  box-shadow: 0 0 15px -2px #292727;
-  border: 1px solid rgb(67, 67, 75);
-  padding: 10px 20px;
-  margin-bottom: 5px;
-  border-radius: 5px;
-}
-
-.ListItem:hover {
-  background-color: #6a9ca8;
-  cursor: pointer;
-}
-
-/* Making ListIten responsiviness */
-@media (max-width: 500px) {
-  .ListItem {
-    color: whitesmoke;
-  }
-}
