+++ src/components/AddTodoForm.module.css
@@ -1,0 +1,12 @@
+/* Styling the form and items */
+.todo {
+  border: 1px solid grey;
+  padding: 10px 20px;
+  margin-bottom: 10px;
+  border-radius: 5px;
+}
+form {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 10px;
+}
